Moved table into cli-lib
