+++ packages/cli-lib/package.json
@@ -23,7 +23,7 @@
     "prettier": "^1.19.1",
     "request": "^2.87.0",
     "request-promise-native": "^1.0.7",
-    "table": "^6.0.3",
+    "table": "^6.0.4",
     "unixify": "1.0.0"
   },
   "engines": {
