+++ packages/cli/commands/config/list.js
@@ -5,6 +5,10 @@ const {
   validateConfig,
   checkAndWarnGitInclusion,
 } = require('@hubspot/cli-lib');
+const {
+  getTableContents,
+  getTableHeader,
+} = require('@hubspot/cli-lib/lib/table');
 
 const {
   addConfigOptions,
@@ -15,7 +19,6 @@ const {
 const { trackCommandUsage } = require('../../lib/usageTracking');
 const { logDebugInfo } = require('../../lib/debugInfo');
 const { validateAccount } = require('../../lib/validation');
-const { getTableContents, getTableHeader } = require('../../lib/table');
 
 const loadAndValidateOptions = async options => {
   setLogLevel(options);
