+++ packages/cli/lib/links.js
@@ -2,9 +2,12 @@ const { getEnv } = require('@hubspot/cli-lib/lib/config');
 const { ENVIRONMENTS } = require('@hubspot/cli-lib/lib/constants');
 const { getHubSpotWebsiteOrigin } = require('@hubspot/cli-lib/lib/urls');
 const { logger } = require('@hubspot/cli-lib/logger');
+const {
+  getTableContents,
+  getTableHeader,
+} = require('@hubspot/cli-lib/lib/table');
 
 const open = require('open');
-const { getTableContents, getTableHeader } = require('./table');
 
 const logSiteLinks = accountId => {
   const linksAsArray = getSiteLinksAsArray(accountId).map(l => [
