+++ packages/cli/package.json
@@ -17,7 +17,6 @@
     "open": "^7.0.3",
     "ora": "^4.0.3",
     "shelljs": "0.8.3",
-    "table": "^6.0.3",
     "update-notifier": "3.0.1",
     "yargs": "15.4.1"
   },
