+++ packages/serverless-dev-runtime/package.json
@@ -14,7 +14,6 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "fs-extra": "^9.0.1",
-    "table": "^6.0.4",
     "tmp": "^0.2.1"
   },
   "publishConfig": {
